Separated code editor files. Fixed dart:mirrors imports
