+++ lib/editor/css/ace_editor_component.css
