+++ lib/editor/html/action_region_editor_component.html
