+++ lib/editor/html/lesson_editor_component.html
