+++ lib/html/lesson_list_component.html
@@ -1,6 +1,6 @@
 <h2>Lesson List</h2>
 <ul>
     <li *ngFor="let lesson of lessons">
-        <a [routerLink]="['Lesson', {'lesson_name': lesson}]">{{lesson}}</a>
+        <a [routerLink]="['Lesson', {'lesson_name': lesson}]">{{lesson}}</a><a [routerLink]="['Lesson Editor', {"
     </li>
 </ul>
